pdf1_0: add Real to GenericObject

+++ src/pdfio/pdf1_0/istream/read_real.cpp
@@ -4,7 +4,7 @@ namespace pdf1_0 = pdfio::pdf1_0;
 
 std::istream &operator>>(std::istream &istream, pdf1_0::Real &real)
 {
-  int buffer;
+  double buffer;
   if(istream >> buffer)
   {
     real = buffer;
